refactoring: moved Neo4jUtil to neo4j-core
