+++ neo4j-core/src/main/java/at/srfg/graphium/neo4j/persistence/Neo4jUtil.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package at.srfg.graphium.mapmatching.neo4j.matcher.impl;
+package at.srfg.graphium.neo4j.persistence;
 
 import org.neo4j.graphdb.traversal.Evaluators;
 import org.neo4j.graphdb.traversal.TraversalDescription;
